Modbus control

+++ main/peripherals/controllo_digitale.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 
-void controllo_digitale_init(void);
-int controllo_digitale_get_signal_on(void);
+void     controllo_digitale_init(void);
+uint8_t  controllo_digitale_get_signal_on(void);
 uint16_t controllo_digitale_get_analog_speed(void);
-int controllo_digitale_get_perc_speed(void);
+int      controllo_digitale_get_perc_speed(void);
 
 #endif
